Security issue fixed.
The files that are publicly accessible are now in the directory "public". All other files are in the folder lib and they are not accessible via http.
This is important because now you don't have access to any php file other than public/index.php and public/setup/install.php.