Added dummy support for several Markup Languages like YAML and JSON. It will be ready to use soon.
I also added a module for basic filesystem interaction, which doesn't work yet either.

Finally I added a markdown readme file and modified my email adress in the header of the files to "p.vorbach@gmail.com". Maybe this will change back in the future, but at the moment I don't want to be addicted to the domain "genitis.org".